Implement Support for Batched Jobs (Coalesce Multiple Enqueues into Larger Jobs)

+++ examples/raw_throughput/active_job_throughput.rb
@@ -11,12 +11,12 @@
 # Change the environment variable JOB_COUNT to control how many jobs are
 # enqueued.
 
-require 'active_job'
-require 'active_job/queue_adapters/zizq_adapter'
-require_relative './setup'
+require "active_job"
+require "active_job/queue_adapters/zizq_adapter"
+require_relative "./setup"
 
-require 'zizq'
-require 'async'
+require "zizq"
+require "async"
 
 # --- Setup ----
 
@@ -33,9 +33,7 @@ class TestJob < ActiveJob::Base
   self.queue_adapter = :zizq
 
   def perform(n)
-    if n == JOB_COUNT
-      Process.kill("TERM", Process.pid)
-    end
+    Process.kill("TERM", Process.pid) if n == JOB_COUNT
   end
 end
 
@@ -45,26 +43,24 @@ enqueue_started_at = Process.clock_gettime(Process::CLOCK_MONOTONIC)
 
 Sync do
   (1..JOB_COUNT).each_slice(1_000) do |nums|
-    ActiveJob.perform_all_later(nums.map {|n| TestJob.new(n)})
+    ActiveJob.perform_all_later(nums.map { |n| TestJob.new(n) })
   end
 end
 
 enqueue_finished_at = Process.clock_gettime(Process::CLOCK_MONOTONIC)
-enqueue_elapsed = Process.clock_gettime(Process::CLOCK_MONOTONIC) - enqueue_started_at
+enqueue_elapsed =
+  Process.clock_gettime(Process::CLOCK_MONOTONIC) - enqueue_started_at
 
 puts format(
-  "Enqueued %d jobs in %.3fs (%.3f jobs/sec).",
-  JOB_COUNT,
-  enqueue_elapsed,
-  JOB_COUNT/enqueue_elapsed
-)
+       "Enqueued %d jobs in %.3fs (%.3f jobs/sec).",
+       JOB_COUNT,
+       enqueue_elapsed,
+       JOB_COUNT / enqueue_elapsed
+     )
 
 # --- Dequeue Phase ----
 
-worker = Zizq::Worker.new(
-  thread_count: THREADS,
-  fiber_count: FIBERS,
-)
+worker = Zizq::Worker.new(thread_count: THREADS, fiber_count: FIBERS)
 
 Signal.trap("TERM") { worker.stop }
 
@@ -73,11 +69,12 @@ dequeue_started_at = Process.clock_gettime(Process::CLOCK_MONOTONIC)
 worker.run
 
 dequeue_finished_at = Process.clock_gettime(Process::CLOCK_MONOTONIC)
-dequeue_elapsed = Process.clock_gettime(Process::CLOCK_MONOTONIC) - dequeue_started_at
+dequeue_elapsed =
+  Process.clock_gettime(Process::CLOCK_MONOTONIC) - dequeue_started_at
 
 puts format(
-  "Dequeued %d jobs in %.3fs (%.3f jobs/sec).",
-  JOB_COUNT,
-  dequeue_elapsed,
-  JOB_COUNT/dequeue_elapsed
-)
+       "Dequeued %d jobs in %.3fs (%.3f jobs/sec).",
+       JOB_COUNT,
+       dequeue_elapsed,
+       JOB_COUNT / dequeue_elapsed
+     )
